refactor /home
